rename main_model_wrapper to main_model
