+++ tests/benchmark_cpp/fictional_grid_generator.hpp
@@ -5,7 +5,7 @@
 #pragma once
 
 #include <power_grid_model/auxiliary/meta_data_gen.hpp>
-#include <power_grid_model/main_model_wrapper.hpp>
+#include <power_grid_model/main_model.hpp>
 
 #include <algorithm>
 #include <random>
